# Improve performance of fiber creation by using pool allocation strategy.