--- a/common.mk
+++ b/common.mk
@@ -924,9 +924,8 @@ strlcpy.$(OBJEXT): {$(VPATH)}strlcpy.c
 strstr.$(OBJEXT): {$(VPATH)}strstr.c
 nt.$(OBJEXT): {$(VPATH)}nt.c
 
-COROUTINE_SRC = $(COROUTINE_H:.h=).$(ASMEXT)
 .coroutine_obj $(COROUTINE_OBJ): \
-	{$(VPATH)}$(COROUTINE_SRC:/ucontext/Context.S=/ucontext/Context.c) \
+	{$(VPATH)}$(COROUTINE_SRC) \
 	$(COROUTINE_H:/Context.h=/.time)
 $(COROUTINE_H:/Context.h=/.time):
 	$(Q) $(MAKEDIRS) $(@D)