--- a/coroutine/arm32/Context.h
+++ b/coroutine/arm32/Context.h
@@ -11,6 +11,7 @@
 #include <string.h>
 
 #define COROUTINE __attribute__((noreturn)) void
+#define COROUTINE_LIMITED_ADDRESS_SPACE
 
 enum {COROUTINE_REGISTERS = 8};