--- a/coroutine/win32/Context.h
+++ b/coroutine/win32/Context.h
@@ -11,6 +11,7 @@
 #include <string.h>
 
 #define COROUTINE __declspec(noreturn) void __fastcall
+#define COROUTINE_LIMITED_ADDRESS_SPACE
 
 /* This doesn't include thread information block */
 enum {COROUTINE_REGISTERS = 4};