https://bugs.ruby-lang.org/issues/15997

It would be good to get some feedback on this PR.

```
$ make benchmark ITEM=fiber COMPARE_RUBY="../build-master/ruby"
Calculating -------------------------------------
                     compare-ruby  built-ruby 
          make_chain       77.840      73.809 i/s -     500.000 times in 6.423414s 6.774286s
        resume_chain      161.783     168.189 i/s -     500.000 times in 3.090556s 2.972852s
  vm2_fiber_allocate      68.114k    112.117k i/s -    200.000k times in 2.936240s 1.783851s
    vm2_fiber_switch       7.812M      8.295M i/s -     10.000M times in 1.280088s 1.205587s

Comparison:
                       make_chain
        compare-ruby:        77.8 i/s 
          built-ruby:        73.8 i/s - 1.05x  slower ** n/c within margin of error

                     resume_chain
          built-ruby:       168.2 i/s 
        compare-ruby:       161.8 i/s - 1.04x  slower ** n/c within margin of error

               vm2_fiber_allocate
          built-ruby:    112117.0 i/s 
        compare-ruby:     68114.3 i/s - 1.65x  slower ** significant improvement

                 vm2_fiber_switch
          built-ruby:   8294714.5 i/s 
        compare-ruby:   7811962.9 i/s - 1.06x  slower ** n/c within margin of error
```