@ioquatix 2019-06-02
I want to test this in real world code too, but I am on holiday I don't have my desktop computer so I can't really test it at the extreme limits.