@ioquatix 2019-06-02
Also, I don't expose `Fiber::Pool` although conceptually it could be exposed pretty easily and I think that would be a good idea.